protocol: move go/protocol => go/protocols/flow
Enable side-by-side addition of future protocols.
The `flow` package is focused on core types and runtime bits.
