+++ go/flow/catalog_test.go
@@ -3,7 +3,7 @@ package flow
 import (
 	"testing"
 
-	pf "github.com/estuary/flow/go/protocol"
+	pf "github.com/estuary/flow/go/protocols/flow"
 	"github.com/stretchr/testify/require"
 	pb "go.gazette.dev/core/broker/protocol"
 )
