+++ go/flow/ingest.go
@@ -7,7 +7,7 @@ import (
 	"sync/atomic"
 	"time"
 
-	pf "github.com/estuary/flow/go/protocol"
+	pf "github.com/estuary/flow/go/protocols/flow"
 	"go.gazette.dev/core/broker/client"
 	pb "go.gazette.dev/core/broker/protocol"
 	"go.gazette.dev/core/message"
