+++ go/ingest/grpc_api_test.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"testing"
 
-	pf "github.com/estuary/flow/go/protocol"
+	pf "github.com/estuary/flow/go/protocols/flow"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/grpc"
 )
