+++ go/ingest/http_api.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"net/http"
 
-	pf "github.com/estuary/flow/go/protocol"
+	pf "github.com/estuary/flow/go/protocols/flow"
 	log "github.com/sirupsen/logrus"
 	pb "go.gazette.dev/core/broker/protocol"
 	"go.gazette.dev/core/broker/protocol/ext"
