+++ go/materialize/config.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"strings"
 
-	pf "github.com/estuary/flow/go/protocol"
+	pf "github.com/estuary/flow/go/protocols/flow"
 	log "github.com/sirupsen/logrus"
 	"go.gazette.dev/core/consumer"
