+++ go/protocols/flow/document_extensions.go
@@ -1,4 +1,4 @@
-package protocol
+package flow
 
 import (
 	"bufio"
