+++ go/protocols/flow/document_extensions_test.go
@@ -1,4 +1,4 @@
-package protocol
+package flow
 
 import (
 	"bufio"
