+++ go/protocols/flow/shuffle_extensions_test.go
@@ -1,4 +1,4 @@
-package protocol
+package flow
 
 import (
 	"testing"
