+++ go/runtime/materialize.go
@@ -8,7 +8,7 @@ import (
 	"github.com/estuary/flow/go/flow"
 	"github.com/estuary/flow/go/labels"
 	"github.com/estuary/flow/go/materialize"
-	pf "github.com/estuary/flow/go/protocol"
+	pf "github.com/estuary/flow/go/protocols/flow"
 	"github.com/estuary/flow/go/shuffle"
 	cache "github.com/hashicorp/golang-lru"
 	log "github.com/sirupsen/logrus"
