+++ go/shuffle/subscriber.go
@@ -7,7 +7,7 @@ import (
 	"math/bits"
 	"sort"
 
-	pf "github.com/estuary/flow/go/protocol"
+	pf "github.com/estuary/flow/go/protocols/flow"
 	pb "go.gazette.dev/core/broker/protocol"
 	"go.gazette.dev/core/message"
 )
