+++ go/shuffle/subscriber_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 	"time"
 
-	pf "github.com/estuary/flow/go/protocol"
+	pf "github.com/estuary/flow/go/protocols/flow"
 	"github.com/stretchr/testify/require"
 	pb "go.gazette.dev/core/broker/protocol"
 	"go.gazette.dev/core/message"
